requirements.txt
dwave-ocean-sdk==2.0.0
penaltymodel-maxgap==0.5.2
pandas==1.0.1
networkx==2.4
